feat(screen): change countdowntimer to today screen
